+++ apps/docs/src/content/03-components/form-controls/checkbox-group/examples/form.tsx
@@ -1,5 +1,4 @@
 import {
-  Button,
   Checkbox,
   CheckboxGroup,
   Label,
@@ -8,6 +7,7 @@ import {
 import { useForm } from "react-hook-form";
 import {
   Form,
+  SubmitButton,
   typedField,
 } from "@mittwald/flow-react-components/react-hook-form";
 import { sleep } from "@/content/03-components/actions/action/examples/lib";
@@ -32,7 +32,7 @@ export default () => {
             <Checkbox value="write">Schreiben</Checkbox>
           </CheckboxGroup>
         </Field>
-        <Button type="submit">Speichern</Button>
+        <SubmitButton>Speichern</SubmitButton>
       </Form>
     </Section>
   );