+++ apps/docs/src/content/03-components/form-controls/combo-box/examples/form.tsx
@@ -1,5 +1,4 @@
 import {
-  Button,
   ComboBox,
   Label,
   Option,
@@ -8,6 +7,7 @@ import {
 import { useForm } from "react-hook-form";
 import {
   Form,
+  SubmitButton,
   typedField,
 } from "@mittwald/flow-react-components/react-hook-form";
 import { sleep } from "@/content/03-components/actions/action/examples/lib";
@@ -31,7 +31,7 @@ export default () => {
             <Option>anotherdomain.com</Option>
           </ComboBox>
         </Field>
-        <Button type="submit">Speichern</Button>
+        <SubmitButton>Speichern</SubmitButton>
       </Form>
     </Section>
   );