+++ apps/docs/src/content/03-components/form-controls/file-field/examples/form.tsx
@@ -9,6 +9,7 @@ import {
 import { useForm } from "react-hook-form";
 import {
   Form,
+  SubmitButton,
   typedField,
 } from "@mittwald/flow-react-components/react-hook-form";
 import { sleep } from "@/content/03-components/actions/action/examples/lib";
@@ -56,7 +57,7 @@ export default () => {
             />
           ))}
         </FileCardList>
-        <Button type="submit">Hochladen</Button>
+        <SubmitButton>Hochladen</SubmitButton>
       </Form>
     </Section>
   );